vp9/encoder/vp9_onyx_if: fix vp9_compute_qdelta_by_rate() warnings
- Make base_q_index and integer.
- Add a cast.

+++ vp9/encoder/vp9_onyx_if.c
@@ -219,7 +219,7 @@ int vp9_compute_qdelta(const VP9_COMP *cpi, double qstart, double qtarget) {
 // to a value that should equate to thegiven rate ratio.
 
 int vp9_compute_qdelta_by_rate(VP9_COMP *cpi,
-                               double base_q_index, double rate_target_ratio) {
+                               int base_q_index, double rate_target_ratio) {
   int i;
   int base_bits_per_mb;
   int target_bits_per_mb;
@@ -233,7 +233,7 @@ int vp9_compute_qdelta_by_rate(VP9_COMP *cpi,
                                         base_q_index, 1.0);
 
   // Find the target bits per mb based on the base value and given ratio.
-  target_bits_per_mb = rate_target_ratio * base_bits_per_mb;
+  target_bits_per_mb = (int)(rate_target_ratio * base_bits_per_mb);
 
   // Convert the q target to an index
   for (i = cpi->rc.best_quality; i < cpi->rc.worst_quality; i++) {
